ui/requirements.txt
# FastAPI Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0

# Database
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.2.0
pyyaml>=6.0.0
apscheduler>=3.10.0

# Optional: AWS integration
boto3>=1.34.0
